controller/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
grpcio==1.38.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
isodate==0.6.0
numpy==1.21.1
packaging==21.2
pandas==1.3.1
pip==20.1.1
pluggy==1.0.0
protobuf==3.17.3
py==1.11.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
rdflib==6.0.2
requests==2.26.0
setuptools==47.1.0
six==1.16.0
soupsieve==2.2.1
SPARQLWrapper==1.8.5
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.6.0
zope.interface==5.4.0
durable-rules==2.0.28